Inline IDC logic with annualizer
